Close #10: Underscore to spaces
In phylogeny it is common to substitute spaces to underscores
to make them compatible with newick format. We remove underscores
from names like

Pseudocercospora_dendrobii

But we co leave underscores if name has even one space

Pseudocercospora dendrobii RNA_033231

is not converted. However names with spaces before or after will be
converted as well --

  Pseudocercospora_dendrobii